Fixes for build errors, support more types of component definitions
- Adds null checks in the logic that traverses the AST.  Prevents users
from getting 'unable to access property 'type' of undefined' errors.
- displayName property definitions are now emitted directly after the
variable / function / class definition, rather than at the end of the
file. Fixes build errors that a few people hit.
- Now detects when component is defined as extending JSXElement and
when it is defined as a class extending Component/PureComponent
- Only parse files that end in .js/.jsx/.ts/.tsx
- Ignore function names / variable names that are lowercase. This is
a simple heuristic to avoid adding displayName to helper functions.