Upgrade plugin to support Webpack 5
- Updates peer dependency to require Webpack 5
- Updates the position at which the `displayName` property
is inserted into the source. For certain cases the previous location we
used was causing us to insert the property before the component itself was defined.
- Updates dependencies in tests to use React 17 and Webpack 5, along
with other dependencies that are compatible with these major versions.
In some cases I had to pull in the source of third party dependencies
directly into the test examples since those dependencies haven't been
updated for React 17 yet.
- Updated unit tests to run sequentially and to pull source from the
file system rather than from Webpack. Webpack no longer contains the
emitted source as a property, which is why I now pull the source from
the file system. But because I rewrite the same source file for each
Babel config, there is a race-condition if I run this in parallel, which
is why I moved it to run sequentially.
- Webpack's tree-shaking algorithm seems to have gotten better in Webpack 5
because it detected instances in my tests where the component is not referenced
and thus kept it out of the emitted source. I had to refactor some of the examples
to make sure the components have references and changed some of the expected
component counts in the unit tests.